refactor client to provider and repository
