+++ rsky-feedgen/diesel.toml
@@ -2,8 +2,8 @@
 # see https://diesel.rs/guides/configuring-diesel-cli
 
 [print_schema]
-file = "src/schema.rs"
 custom_type_derives = ["diesel::query_builder::QueryId"]
+file                = "src/schema.rs"
 
 [migrations_directory]
 dir = "migrations"
