+++ rsky-pds/diesel.toml
@@ -2,9 +2,9 @@
 # see https://diesel.rs/guides/configuring-diesel-cli
 
 [print_schema]
-file = "src/schema.rs"
 custom_type_derives = ["diesel::query_builder::QueryId"]
-schema = "pds"
+file                = "src/schema.rs"
+schema              = "pds"
 
 [migrations_directory]
 dir = "migrations"
