+++ rsky-pds/src/apis/firefly/mod.rs
@@ -1,4 +1,4 @@
-mod client;
+pub mod providers;
 mod contracts;
 pub mod create_transfer_request;
 pub mod fulfill_transfer_request;
@@ -7,3 +7,4 @@ pub mod get_wallet_state_and_history;
 pub mod models;
 mod state;
 pub mod transfer;
+mod repositories;
