+++ rsky-pds/src/apis/firefly/transfer.rs
@@ -1,9 +1,8 @@
-use rocket::serde::json::Json;
-use crate::apis::firefly::client::FireflyClient;
+use crate::apis::firefly::providers::FireflyProvider;
 use crate::apis::ApiError;
 use crate::config::ServerConfig;
+use rocket::serde::json::Json;
 use rocket::State;
-
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct TransferRequest {
@@ -16,14 +15,14 @@ pub struct TransferRequest {
 #[rocket::post("/transfer", format = "json", data = "<body>")]
 pub async fn transfer(
     body: Json<TransferRequest>,
-    cfg: &State<ServerConfig>,
+    provider: &State<FireflyProvider>,
     // auth: AccessStandard  // remove comment to turn on auth
     // TODO: auth should be turned on when we'll solve storage parameters per user
 ) -> Result<(), ApiError> {
     let TransferRequest {
         amount, to_address, ..
     } = body.into_inner();
-    let client = FireflyClient::new(&cfg)?;
+    let client = provider.firefly();
     let _response_hash = client.transfer_request(&to_address, amount).await?;
 
     Ok(())
